refactor: to make tests for #8 easier to grasp
