+++ Chat.Tests/Logic/IChatHubAssertions.cs
@@ -1,0 +1,11 @@
+namespace Chat.Tests.Logic
+{
+    public interface IChatHubAssertions
+    {
+        ClientOperationsSpy Caller { get; }
+
+        ClientOperationsSpy All { get; }
+
+        ClientOperationsSpy Others { get; }
+    }
+}
