+++ Chat.Tests/Web/WebBrowserTests.cs
@@ -1,16 +1,15 @@
-﻿namespace Chat.Tests
+﻿namespace Chat.Tests.Web
 {
     using System.Collections.Generic;
 
     using Chat.Logic;
 
-    using Coypu.Matchers;
-
     using FluentAssertions;
+
     using Xunit;
 
     public class WebBrowserTests : IClassFixture<WebServerFixture>
-    { 
+    {
         private readonly IChatConfiguration configuration = new ChatConfiguration();
 
         [Fact]
@@ -134,7 +133,7 @@
                 alice.Join();
                 bob.Join();
                 alice.SendMessage("Beware Bob! I want to <script>alert('hack you!');</script>");
-                
+
                 // checking against dialog would be problematic in phantomjs.
                 // since it does not support alert directly webdriver's hasdialog won't work.
                 // instead we just explicitly validate encoded version appears on screen.
